% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{assignGammas}
\alias{assignGammas}
\title{Assigns each sample to an LPD group according to the gamme value}
\usage{
assignGammas(project, automataID, genes)
}
\arguments{
\item{project}{.}

\item{automataID}{.}
}
\description{
The title is quite selfexplanatory, i am so sleepy.
}
