% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{calculateCorrelation}
\alias{calculateCorrelation}
\title{Calculates correlation and chooses best combination}
\usage{
calculateCorrelation(project, automataID, genes)
}
\arguments{
\item{project}{.}

\item{automataID}{.}
}
\description{
Calculates the correlation for each combination and selects the one with closest Pearson to 0 and
p-value lower than 0.05 as the best.
}
