% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgSites.R
\docType{data}
\name{cgSites}
\alias{cgSites}
\title{cg sites for methylation}
\format{
CSV file
}
\usage{
data(cgSites)
}
\description{
Data downloaded from the TCGA containing the info for each cg site such as chromosome, start and end,
gene symbol, gene type, ENSEMBL transcript ID, Position to TSS, CGI coordinates and feature type.
}
\keyword{datasets}
