% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{cleanMatrix}
\alias{cleanMatrix}
\title{Function to remove the duplicated plates}
\usage{
cleanMatrix(project, automataID, microarray = FALSE)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{microarray}{Wether the expression data comes from RNA-seq counts or microarray levels}
}
\value{
A matrix without duplicates
}
\description{
Removes duplicated samples with different plates or/and vials.
}
