% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{compareGammas}
\alias{compareGammas}
\title{Compare gammas to select the best run of each combination}
\usage{
compareGammas(project, automataID)
}
\description{
Checks each one of the 100 gamma files and calculates a mean gamma. The file with the lowest distance
between the mean and its values is chosen as the most representative.
}
