% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{createIntersection}
\alias{createIntersection}
\title{Does the intersection}
\usage{
createIntersection(project, automataID)
}
\description{
Checks which genes are differentially expressed, differentially methylated, have mutations
and copy number variations.
}
