% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarcodes.R
\name{createTCGAfolder}
\alias{createTCGAfolder}
\title{Create TCGA_results folder}
\usage{
createTCGAfolder(project, automataID)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}
}
\value{
Generates a series of folders and subfolders for each cancer project and automata ID.
}
\description{
Creates the folder TCGA_results to save the data generated by different functions of Automata.
}
