% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{diffExpGenes}
\alias{diffExpGenes}
\title{Get the differentially expressed genes}
\usage{
diffExpGenes(project, automataID, microarray = FALSE)
}
\arguments{
\item{project}{.}

\item{automataID}{.}

\item{microarray}{Defaults to FALSE}
}
\description{
Creates CSV files with the log2Folf of the top 500 genes and plots it in a tornado plot for each LPD process.
}
