% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadClinical}
\alias{downloadClinical}
\title{Download clinical data from the TCGA}
\usage{
downloadClinical(project, barcodes, automataID)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{barcodes}{Vector containing participant barcodes for the cases to download.}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}
}
\value{
Downloads a CSV file per sample containing the clinical data and creates a merged CSV with all
a sample in each column.
}
\description{
Downloads clinical data for the specified project in the TCGA.
}
\note{
Barcodes need to contain the participant code (eg. TCGA-WS-AA2X). An easy way to get barcodes
is using the function getBarcodes.
}
