% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadCNV}
\alias{downloadCNV}
\title{Download copy number variation (CNV) data from the TCGA}
\usage{
downloadCNV(project, barcodes, automataID)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{barcodes}{Vector containing sample barcodes for the cases to download.}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}
}
\value{
Downloads a CSV file per sample containing the CNV data and creates a merged CSV with all
a sample in each column.
}
\description{
Downloads CNV data for the specified project in the TCGA.
}
\note{
Barcodes need to contain the sample code (01A or 11A among others). An easy way to get barcodes
is using the function getBarcodes.
}
\examples{
downloadCNV(project = "TCGA-CHOL", barcodes = c("TCGA-W5-AA2X-01A", "TCGA-ZU-A8S4-11A"))
}
