% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadExpression}
\alias{downloadExpression}
\title{Download gene expression data from the TCGA}
\usage{
downloadExpression(project, barcodes, automataID, microarray = FALSE)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{barcodes}{Vector containing the sample barcodes to download.}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{microarray}{Wether to check for available microarray data (in the TCGA legacy) or RNA-seq
counts (in the harmonised TCGA). Defaults to FALSE.}
}
\value{
Downloads a CSV file per sample containing the expression data and creates a merged CSV with all
a sample in each column.
}
\description{
Downloads expression data (RNA-seq or microarray) for the specified project in the TCGA.
}
\note{
Barcodes need to contain the sample code (01A or 11A among others). An easy way to get barcodes
is using the function getBarcodes.
}
\examples{
downloadExpression(project = "TCGA-CHOL", barcodes = c("TCGA-W5-AA2X-01A", "TCGA-ZU-A8S4-11A"),
microarray = FALSE)
}
