% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadMeth}
\alias{downloadMeth}
\title{Download methylation data from the TCGA}
\usage{
downloadMeth(project, barcodes, automataID)
}
\arguments{
\item{project}{Name of the TCGA project data to download (eg. TCGA-PRAD or TCGA-BRCA).}

\item{barcodes}{Vector containing sample barcodes for the cases to download.}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}
}
\value{
Downloads all the raw data in a folder and creates a  metylation matrix saved in TCGA-results
}
\description{
Downloads metylation data for the specified project in the TCGA.
}
\note{
Barcodes need to contain the sample code (01A or 11A among others). An easy way to get barcodes
is using the function getBarcodes.
}
\examples{
downloadMeth(project = "TCGA-CHOL", barcodes = c("TCGA-W5-AA2X-01A", "TCGA-ZU-A8S4-11A"))
}
