% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadSNV}
\alias{downloadSNV}
\title{Download SNV data from the TCGA}
\usage{
downloadSNV(project, barcodes, automataID)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{barcodes}{Vector containing sample barcodes for the cases to download.}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}
}
\value{
Downloads the four MAF files in the GDC folder and the processed data in TCGA_results.
}
\description{
Downloads a MAF file for each workflow (MuSE, MuTect2, VarScan2, SomaticSniper) and merges them
together keeping the SNPs present in the four of them and the indels present in two of them.
}
\examples{
downloadSNV(project = "TCGA-CHOL", barcodes = c("TCGA-W5-AA2X-01A", "TCGA-ZU-A8S4-11A"))
}
