% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPD.R
\name{estimateParameters}
\alias{estimateParameters}
\title{Estimate best number of processes and sigma value}
\usage{
estimateParameters(project, automataID, path = getwd())
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{path}{Path of the working directory}
}
\description{
Creates a file with the best combinations and plots the likelihood for each one
}
