% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraClinical.R
\name{extraClinical}
\alias{extraClinical}
\title{Extra clinical analysis}
\usage{
extraClinical(project, automataID, clinicalData, process)
}
\arguments{
\item{project}{.}

\item{automataID}{.}

\item{clinicalData}{Import the clinical data from the main clinical analysis function}

\item{process}{LPD process}
}
\description{
Performs further analysis for some specific cancers, so far it's only available
for TCGA-PRAD.
}
