% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvAnalysis.R
\name{generateAberrations}
\alias{generateAberrations}
\title{Generates CNV aberrations per gene}
\usage{
generateAberrations(project, automataID)
}
\arguments{
\item{project}{.}

\item{automataID}{.}
}
\description{
Creates genomic ranges for the CNV data downloaded from the TCGA and finds overlaps with genomic ranges
originated from ENSEMBL.
}
