% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPD.R
\name{generateJobFolder}
\alias{generateJobFolder}
\title{Creates job folders}
\usage{
generateJobFolder(project, automataID, path, sigma, lpdStage, n_process = NULL)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{path}{Filepath to the original working directory to return after finishing the function.}

\item{sigma}{Sigma value (function to run inside of a loop)}
}
\description{
Generates a folder with the required data to run LPD and changes the working directory to it.
}
