% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generates report to check results}
\usage{
generateReport(project, automataID, microarray, workingPath = getwd())
}
\description{
Creates and HTML file containing the resuls from the whole automata pipeline
}
