% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesRanges.R
\docType{data}
\name{genesRanges}
\alias{genesRanges}
\title{genes ranges for cnv}
\format{
CSV file
}
\usage{
data(genesRanges)
}
\description{
Data downloaded from biomart containing the range for each gene in the human genome.
}
\keyword{datasets}
