% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarcodes.R
\name{getBarcodes}
\alias{getBarcodes}
\title{Get barcodes from the TCGA}
\usage{
getBarcodes(project, automataID, microarray = FALSE)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{microarray}{Wether to check for available microarray data (in the TCGA legacy) or RNA-seq
counts (in the harmonised TCGA). Defaults to FALSE.}
}
\value{
A CSV file containing the participant and sample barcode, the sample typem the workflow (RNA-seq)
or microarray), the project and the assigned automata id.
}
\description{
For a given TCGA cancer project, fetches the cases with available expression data and returns the
barcodes assigned to those cases.
}
