% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{getQueryFilePaths}
\alias{getQueryFilePaths}
\title{Get filepaths from GDCquery}
\usage{
getQueryFilePaths(query)
}
\arguments{
\item{query}{GDCquery object generated by TCGAbiolinks::GDCquery}
}
\value{
Character vector of the pathway to the GDCfolder
}
\description{
Gets the filepath to which the raw data will be downloaded from the GDCquery created
by TCGAbiolinks.
}
