% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPD.R
\name{launchLPD_B_ADA}
\alias{launchLPD_B_ADA}
\title{Runs the second stage of LPD (AADA)}
\usage{
launchLPD_B_ADA(project, automataID, n_process, sigma, path = getwd())
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{n_process}{Number of processes estimated as optimal}

\item{sigma}{Sigma value estimated as optimal.}

\item{path}{Pathway to TCGA_results folder. Defaults to working directory.}
}
\description{
Launches 90 jobs trying different sigma values with different number of processes. COmpatible with ADA.
}
