% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARTannotations.R
\docType{data}
\name{MARTannotations}
\alias{MARTannotations}
\title{mart annotations to change between ensembl and hgnc symbol}
\format{
CSV file
}
\usage{
data(MARTannotations)
}
\description{
Put more info here.
}
\keyword{datasets}
