% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{methylationAnalysis}
\alias{methylationAnalysis}
\title{Analysis of methylation data}
\usage{
methylationAnalysis(project, automataID)
}
\arguments{
\item{project}{.}

\item{automataID}{.}
}
\description{
Performs an analysis of the methylation data downloaded from the TCGA
}
