% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normaliseExpression}
\alias{normaliseExpression}
\title{Normalise expression matrix}
\usage{
normaliseExpression(project, automataID, microarray = FALSE)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{microarray}{Wether the expression data comes from RNA-seq counts or microarray levels}
}
\value{
A dataframe with samples in the colums and gene codes in the rows containing the
normalised expression count values
}
\description{
Applies DESeq2 to an RNA-seq / microarray expression matrix.
}
\examples{
normaliseExpression(expressionMatrix)
}
