% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Postprocessing}
\usage{
postprocess(project, automataID)
}
\arguments{
\item{automataID}{}
}
\description{
Process each run and select the best one across the three combinationss and assigns each sample
to the LPD group to which they are more represented.
}
