% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{runDendrogram}
\alias{runDendrogram}
\title{Runs a dendrogram}
\usage{
runDendrogram(project, automataID)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}
}
\description{
Runs a dendogram to compare its results with the LPD clustering using the top 500 genes per sample
}
