% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPD.R
\name{runLPD_A_ADA}
\alias{runLPD_A_ADA}
\title{Runs the first stage of LPD (ada)}
\usage{
runLPD_A_ADA(project, automataID, path = getwd())
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{path}{Pathway to TCGA_results folder. Defaults to working directory.}
}
\description{
Launches 90 jobs trying different sigma values with different number of processes for ADA.
}
