% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPD.R
\name{runLPD_B_ADA}
\alias{runLPD_B_ADA}
\title{Launches jobs with best combinations for LPD-B (ada)}
\usage{
runLPD_B_ADA(project, automataID, path = getwd())
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{path}{Pathway to TCGA_results folder. Defaults to working directory.}

\item{n_process}{Number of processes estimated as optimal}

\item{sigma}{Sigma value estimated as optimal.}
}
\description{
Launches 90 jobs trying different sigma values with different number of processes.
}
