% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{runPathwayAnalysis}
\alias{runPathwayAnalysis}
\title{Runs the pathway analysis for the given vector names}
\usage{
runPathwayAnalysis(
  project,
  automataID,
  process,
  geneNames,
  contextNames,
  positive.l2FC,
  type
)
}
\arguments{
\item{project}{.}

\item{automataID}{.}

\item{process}{.}

\item{geneNames}{Genes names to analyse}

\item{contextNames}{Gene names of the whole dataset (extracted from TCGA)}

\item{positive.l2FC}{TRUE/FALSE}

\item{type}{Possible options: "expression" ...}
}
\description{
Performs the pathway analysis of the given vector names.
}
