% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvAnalysis.R
\name{snvAnalysis}
\alias{snvAnalysis}
\title{SNV analysis}
\usage{
snvAnalysis(project, automataID)
}
\arguments{
\item{project}{.}

\item{automataID}{.}
}
\description{
Analyse the SNV data downloaded from the TCGA.
}
