% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticSignatures.R
\name{somaticSignaturesAnalysis}
\alias{somaticSignaturesAnalysis}
\title{Analysis of somatic signatures}
\usage{
somaticSignaturesAnalysis(project, automataID)
}
\description{
Does an analysis of somatic signatures
}
