% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{sortGamma}
\alias{sortGamma}
\title{Sets the same order of gamma values columns in all the runs}
\usage{
sortGamma(project, automataID)
}
\description{
Takes the first run of a combination and sorts the rest of the runs according to that one. In order
to identify the corresponding column, Spearman's correlation is used.
}
