% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{topGenes}
\alias{topGenes}
\title{Select the genes with the highest variance}
\usage{
topGenes(project, automataID, microarray = FALSE, uniqueGenes = 500)
}
\arguments{
\item{project}{Name of a TCGA project (eg. TCGA-PRAD or TCGA-BRCA).}

\item{automataID}{ID assigned to the analysis so the files from different analysis are not mixed.}

\item{uniqueGenes}{Number of desired unique genes in the output.}
}
\value{
Matrix containing the columns with the highest variance. The number of columns depends on the
number of uniqueGenes.
}
\description{
Order the expression matrix according to the variance of each column and takes the minimum
number of needed columns to reach the given number of unique genes.
}
